% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscatdif.R
\name{mscatdif}
\alias{mscatdif}
\title{Multistage CAT DIF Function}
\usage{
mscatdif(scored.resp, person.ability, item_difficulty, groupinfo,
  which.focal = NULL, which.ref = NULL, item_fit = NULL,
  item_names = NULL, outfilename = NULL, purification = FALSE,
  FieldTest = FALSE, FTnSize = NULL)
}
\arguments{
\item{x}{is the person data frame containing a column of individual IDs, MLE ability estimates, and scored responses.}

\item{y}{is the item data frame containing item difficulties.}

\item{outsave}{default is FALSE. Save a csv file of the output}

\item{path}{saves the output to specified location when outsave is TRUE.}
}
\description{
This function performs DIF analysis on items from a multi-stage CAT
}
\examples{
set.seed(1010)
ItemData = matrix(sample(rep(0:1, each = 1000),1000),100,10)
PersMeas = rnorm(100, 0, 1)
DemogGroup = sample(rep(c("O","H"), each = 100), 100)
ItemMeas = rnorm(10, 0, 1)
mscatdif(ItemData, PersMeas, ItemMeas, DemogGroup)
}
\keyword{CAT}
\keyword{DIF}
\keyword{multistage}
