% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscatdif.R
\name{mscatdif}
\alias{mscatdif}
\title{Multistage CAT DIF Function}
\usage{
mscatdif(scored.resp, person.ability, item.difficulty, group.info,
  which.focal = NULL, which.ref = NULL, item.fit = NULL,
  item.names = NULL, outfilename = NULL, purification = FALSE,
  field.test = FALSE, FTnSize = NULL)
}
\arguments{
\item{scored.resp}{is a matrix of scored responses.}

\item{person.ability}{is a vector of person ability estimates.}

\item{item.difficulty}{is a vector of item difficulty measures.}

\item{group.info}{is a vector of person demographic variable.}

\item{which.focal}{is the desired focal group. Random group chosen if NULL.}

\item{which.ref}{is the desired reference group. Random group chosen if NULL.}

\item{item.fit}{is a vector of item fit information. Default = NULL.}

\item{item.names}{is a vector of item names. Default = NULL.}

\item{outfilename}{is the name of the desired .csv output file. Default = NULL.}

\item{purification}{run purification? Default = FALSE.}

\item{field.test}{is it a field test? Default = FALSE.}

\item{FTnSize}{if !is.null(field.test) then FTnSize integer scalar must be specified.}
}
\description{
This function performs DIF analysis on items from a multi-stage CAT
}
\examples{
set.seed(1010)
ItemData = matrix(sample(rep(0:1, each = 1000),1000),100,10)
PersMeas = rnorm(100, 0, 1)
DemogGroup = sample(rep(c("O","H"), each = 100), 100)
ItemMeas = rnorm(10, 0, 1)
mscatdif(ItemData, PersMeas, ItemMeas, DemogGroup)
}
\keyword{CAT}
\keyword{DIF}
\keyword{multistage}
